Updated Made to address transitions
Now there is a seperate function that remove the strides at the transitions between targets.  This is important that the events not be removed so that the python and nexus strides are similar for cross correlation purposes, and that the labtools params and the python params align. 

I also made the plotting parts of analyze different methods so that the analysis part of the code is less cluttered.