Commented error-throwing on strokeSubjectData constructor
For some reason old files don't have a valid 'affected' side saved for
stroke subjects ('R' and 'L' are the valid values). This makes the
constructor throw an error everytime such a file is loaded. Normally it
isn't an issue, because there seems to be a try/catch structure dealing
with the issue (not sure where). But when dbstop if error is enabled, it
will stop on that line for every load, which is bothersome. Since the
error had no practical effect anyway, I commented it.