(Potential) behavior change requesting more steps than available in adaptationData
IMPORTANT: this commit does NOT change code behavior, just inserted
'silent' code we may want to use in the future.
Previously, requesting N steps of data from adaptationData if less than
N where available would throw an error, because getEarlyLateIdxs treats
that case by issuing a warning and returning some NaN values, but NaN
values can't be used to extract data.
With the new code (which is NOT enabled yet), the default behavior would
change to throwing a warning, getting the data for the strides that are
available and returning NaN for the rest.