Added monoLS fit
Originally, this was thought as a smoothing with median-filter
qualities: it converges to a monotonic function.
However, it was posed as a constrained LS-fit (where the first
derivative of the signal is constrained to be of fixed sign), which
makes it much more efficient.
Then I generalized it to be able to force the first N derivatives to be
fixed sign.