Fixed issue in computeForceParams
Handrail data is sometimes not stored under HF. Now allowing it to read
XF if HF doesnt exist, and otherwise assigning NaN. On either
contingency, we issue a warning.
Also fixed a mistaken if condition in calcPArameters that generated an
unnecessary warning to be issued every time the function was run (no
practical effect).