GroupDemographics
This will take inputs of the class group adaptation data and will tell you the number of subjects, the mean and standard deviation of age, and the number of males.  For those of you not named Pablo, you can take the output of makeSmatrix and uiCreateStudy and run each group through this one at a time.