Added function to incorporate dependent parameters from file
The idea is that some parameters can be computed as dependent on other parameters, which can be used to avoid storing an ever-growing set of parameters, and only compute them on the fly if necessary. This would also simplify the computeParameters files.
Also changed addNewParameter to call on labTS.addNewParameter