Incorporated btk as part of labtools
Biomechanics toolkit (btk) is an external software package we rely on to import .c3d files from vicon. This change allows us to keep a copy within our code, to avoid having to fetch it externally each time. Should revise the licensing carefully to confirm that this is permitted (to the best of my knowledge it is).