Improved paramSeries.addStrides()
PReviously it was not allowed to addStrides with diff. number of parameters. This is now allowed (with a warning) by padding missing parameters as NaN. This is useful when some trials are missing EMG, angle or other specific data which is present in other trials (by default, neither EMG nor angle parameters are computed for trials with missing data, to allow for them not being present at all in the experiment)