No Positional Bias Removal
This code no longer uses a positional bias removal technique (i.e., OG bias is dependent on the physical position in the lab).  Rather TM and OG and any trial types are treated the same. The bias is computed as the last 40-5 strides of baseline and this is subtracted from all trials of the same type.