Reduce/Clean your data to make it public-safe
By Reducing/Cleaning data, I mean that it removes extra parameters. 
It does this in one of two ways (1) preferred- which requires more recent copies of the code or (2) non-preferred- NaNs out all the parameters you don't want others to see, may need to do this if your code is too old.