Adding Datlogs from Perceptual trials raw, exp, param files
Changed the GUI such that there is an option for the perceptual studies (2AFC task). It shows as box to select when you are using perceptual tasks and it will upload the datlogs to the Metadata in each of the output files.

Important: if you are using the percpetual tasks indicator, make sure to have a folder named "datlogs" where the c3d files are located. It will use this same path to load them. Also, each datlog must be named after each trial it belongs to (same as the nexus files; e.g., Trial01, Trial01, ...., TrialXX)