Improve H-reflex script user input retrieval
Update to always show a pop-up to confirm the participant/session ID. Add a path delimiter to the figure folder path. Add more explicit instructions to the user input prompt (by SL). Save the configuration for the trial number if it does not already exist. Update the H-wave window to 25 - 45 ms after the stimulation artifact peak. Fix typo bugs. Add trial number to figure filename to ensure not overwritten. Update 'tiledlayout' to work for this version of MATLAB by inputting the dimensions with or without the force traces (i.e., 'vertical' input does not work in this version, 2021a). Add plot for line connecting mean values (rather than the fit line only). NWB