Modification Datalog Synchronization (sampling frequency) (#153)
* Experiment Description for Weber Faster

* Fix mistake from Perceptual cues

I was saving the start cues as the stop cues

* Add fast leg option on the GUI

Add the pop up menu for fast leg option.

* Update in datalog synchronization

Before we had hardcoded the sampling frequency. We noticed that the sampling frequency for the GRF data varied depending on the experimental setup. Now this value is not hardcoded but comes from the data.