Modification on reference leg in the code (#155)
* Experiment Description for Weber Faster

* Fix mistake from Perceptual cues

I was saving the start cues as the stop cues

* Add fast leg option on the GUI

Add the pop up menu for fast leg option.

* Update in datalog synchronization

Before we had hardcoded the sampling frequency. We noticed that the sampling frequency for the GRF data varied depending on the experimental setup. Now this value is not hardcoded but comes from the data.

* Add Fast Leg in the c3d2mat GUI

I needed to change some files to actually save the value of the leg which we want to make the fast, such that it would be backward compatible for the info files where do not have this field.

- for the stroke, it will choose the non-affected side to populate the fast left field in the GUI
- for the healthy adults (young and old) will grab the dominant leg information as the fast leg

* Modification on reference leg in the code

The code was using the dominant leg in the GUI or affected side as the reference leg for future processing of the data after c3d2mat. Currently, the reference leg in the code will be assigned to be the slow leg with the new field that was added on c3d2mat GUI.