Updates for Spinal Adaptation project (#168)
* Update parameter normalization

Update 'plotAvgTimeCourse' parameter normalization (i.e., 'removeBiasFlag = 2') to use the last 100 good strides (excluding the last five and first ten strides) in the case that one of the parameters contains is for the H-reflex so that 10 stimuli are included in the average. Remove an unecessary semicolon. NWB

* Add conditional check

Add conditional statement to check that 'GRFData' is not an empty object before evaluating one of its fields. NWB

* Update legend location

Update 'plotAvgTimeCourse' legend code to use the 'best' location and prevent autoupdating. NWB

* Handle bin widths greater than number of strides

Add code block to handle bin widths greater than the number of strides within a trial or condition (previously caused a cryptic error and now only outputs a warning) so that the bin width can be set high enough for three stimuli for Spinal Adaptation project even though there are conditions with ten or fewer strides. NWB