Bug fix for H-reflex (#177)
* Handle stim errors

Handle errors with there being different numbers of stimuli than strides (possibly due to issues with trial splitting by condition) by identifying any stim that should be discarded based on the time of the heel strike events of the corresponding leg. Add warning messages so that user is aware of the discarded stimuli. Remove unnecessary commented code block. Ensure that the window to search for the stimulation artifact and to compute the H-reflex parameters does not exceed the last index of the raw EMG signal (leading to array indexing error). Handle case of H-reflex window of interest being empty. NWB