Tweak H-reflex alignment (#178)
* Update peak finding

Update TA stimulation artifact peak finding algorithm to use the second tallest peak if it exceeds the 1mV threshold and occurs earlier within the window (after observing that the M-like wave peak can be greater in the TAP than the stimulation artifact peak for a small percentage of stimuli in select participants typically with weaker H-reflexes). Update comments and remove irrelevant 'TODO' notes. Fix bug in GenerateHreflexRecruitmentCurvesDuringExp script due to incorrectly passing 'indsStimArtifact' as an input argument (which only exists for some of the calibration experimental conditions). NWB