Fix bug
Handle case of marker trajectory gap occurring at start of end of trial such that there is no frame with the marker before the start or after the end of the gap, which is required for computing the scaling between the reference and the target marker. Update variable names for consistency. NWB