Fix scaling and offset bug
Fix bug in the indexing used to compute the scaling and offset between the reference marker and the target (i.e., previously used the reference marker first and last index of the gap and the target marker index before the gap and after the gap, so one off). Update to only save the trial if changes have been made to save time. Update comments. NWB