Remove unnecessary conditional
Remove unnecessary condition for finding the index of the tibialis anterior stimulation artifact peak to simplify code. Remove check that an index was found and replace call to 'find' with the minimum of the absolute difference even though the times should always be an exact match for improved robustness. Update comments for clarity. Rename variable for consistent naming convention. NWB