Remove wave durations output
Remove the storage and output of the M- and H-wave durations from the 'computeAmplitudes' function since do not see this being a useful parameter to explore outside of this function in the future. Update comments and function documentation to remove references to duration or background EMG (since previously removed). NWB