Reduce save frequency
Add optional input argument boolean to indicate whether or not to save the trial to the functions to process an individual Vicon trial. Update input argument checks and function documentation. Update function to process and fill all marker gaps in a session to prevent functions from saving the trial and instead add code block to save the trial at the end of processing. NWB