Update H-reflex parameter names
Update names of H-reflex parameters for clarity to explicitly mention whether it is an amplitude measure and the muscle for which it is computed. Update to accommodate recent modifications to the H-reflex namespace functions. Handle case of no stimulation artifact indices for a given trial by leaving parameters as default values and returning from the function. Immediately convert amplitudes from 'Volts' to 'mV'. Update comments. NWB