Modify background EMG extraction input
Update 'extractBackgroundEMG' function to accept an array of the number of samples before the indices to begin the window for each leg rather than just a scalar value so that the number of samples before the index can vary on a stride-by-stride basis. Update comments and formatting for clarity. Remove unnecessary 'TODO' note. NWB