Add M-wave labels
Fix function call syntax bug in script to generate H-reflex recruitment curves. Increase M-wave recruitment curve R^2 threshold from 0.5 to 0.8 to reflect the typically higher quality fits observed for this curve relative to the H-wave and the importance of a good fit since making a decision as to the stimulation current based on this curve. Turn off legend border. Add lines and text for M-wave amplitude at optimal stimulation current based on second derivative of the recruitment curve (i.e., point at which M-wave recruitment curve slope is changing the fastest). Add text to indicate the Hmax/Mmax ratio as well as the proportion of Mmax of the M-wave at the selection stimulation current. NWB