Fix bugs
Fix bug due to missing 'MinSeparation' field in updating structure. Fix bug due to setting the minimum number of cameras to continue a trajectory to '1', which is not a possible value in Vicon Nexus leading to Nexus not running the 'Reconstruct And Label Test' pipeline but proceeding as though it had successfully been run. Update possible environmental drift tolerance parameter range to reduce total number of parameter sets. NWB