Updated Process and Fill Marker Gap Session
Added modified akima spline fill for the small gaps that remain after pattern fill. Updated the timeout length to 1200 for all pipelines to help ensure the pipeline runs to completion. For C3 trials, added a section to add red rectangles to highlight turning points (not for other studies). Increase the width of the y-axis when plotting trajectories.

SB 9/15/2025