close figures and remove aux variables to improve memory performance during c3d2mat
There seems to be a memory issue since summer 2025. During c3d2mat, the PC will run out of memory
which is shown as OutOfMemory, OutOfHeapSpace, or png file failed to write rrors.
A better solution is needed to identify why we are running out of memory or do we have a memory leak
Since we don't know the cause now, will try close the figures and remove variables to make the code run for now.