Close figures and delete aux variables to improve memory performance in c3d2mat, bug fix to pass trialtype when getting partial data as parameter series (#201)
* close figures and remove aux variables to improve memory performance during c3d2mat

There seems to be a memory issue since summer 2025. During c3d2mat, the PC will run out of memory
which is shown as OutOfMemory, OutOfHeapSpace, or png file failed to write rrors.
A better solution is needed to identify why we are running out of memory or do we have a memory leak
Since we don't know the cause now, will try close the figures and remove variables to make the code run for now.

* pass the trialtype in the reduce call

This is likely a bug fix, there is no reason to drop the trialtypes, when available, should pass it and preserve it when reducing the data.