Move unused functions to 'deprecated'
Move 'experimentalDetails' function, which has comment at the top of the file explicitly stating to no longer use, into a new 'deprecated' folder to avoid confusion and better organize the code. Move 'getLabelsFromc3d' function, which an exhaustive search revealed as not being called from any piece of code in the 'labTools' repository, into the 'deprecated' folder. NOTE that these files can be easily moved out of the 'deprecated' folder if they return to use and this location causes issues in the ability to call the functions. NWB