Move unused code to 'deprecated'
Move functions that are not called anywhere else in the 'labTools' repository into the 'deprecated' folder to avoid confusion and better organize the code. NOTE that these files can easily be moved out of the 'deprecated' folder if they return to use and this location causes issues in the ability to call the functions. NWB