Add median bias removal option (#205)
* Update plotAvgTimeCourse.m

Add patches for perceptual trials when existing

* Updates for Perceptual Trials Processing

I have changed the description for the Perceptual adaptation S2, also I changed the task to exclude the 3 strides ramp down after the perceptual trials

* Fix non Perceptual tasks data

One of the new parameters for perceptual task ('SLAnotPercTask') was only working properly for the trials that had perceptual tasks interleaved with it, now it works for all instead of population them with NaNs

* Added backward configuration to the GUI

Labtools currently assume that the right belt in the treadmill, is the right belt while in the forward configuration. For the perceptual tasks when we use the backward configuration, this will flip the gait event which is problematic. This configuration setup in the gui should fix that.

* Revert "improve robustness of seperate conditions (#187)"

This reverts commit b53216dde999cf32dec3184ba2470b90dd0fe4fb.

* Modify plotting functions

Changes plotAvgTimeCourses to allow for aligning the perceptual task data. It should not influence the general function unless you have perceptual tasks. Also a function to align the perceptual task data was included.

* Threshold 10N for event detection

Changed by mistake in a previous commit